refactor(cmd): move all cmd to be structured modules
