+++ src/lib.rs
@@ -1,12 +1,17 @@
+pub mod allowance;
+pub mod approve;
 pub mod asset;
 pub mod balances;
 pub mod cmd;
 pub mod config;
 pub mod quote;
 pub mod rebalancer;
+pub mod swap;
 pub mod telemetry;
+pub mod track;
 pub mod unwrap;
 pub mod utils;
+pub mod withdraw;
 pub mod wrap;
 
 pub const APP_NAME: &str = "mfm";
